pyproject.toml
[project]
name = "dermalert-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "drf-nested-routers>=0.94.2",
    "drf-yasg>=1.21.10",
    "pillow>=11.2.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.2.0",
    "ruff>=0.11.4",
]
prod = [
    "gunicorn>=23.0.0",
]
